Update as per Pr feedback to move Dto to same level as API package
