+++ extensions/registration-service/src/test/java/org/eclipse/dataspaceconnector/registration/store/InMemoryParticipantStoreTest.java
@@ -18,7 +18,7 @@ import org.eclipse.dataspaceconnector.registration.authority.model.Participant;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.eclipse.dataspaceconnector.registration.TestUtils.createParticipant;
+import static org.eclipse.dataspaceconnector.registration.authority.TestUtils.createParticipant;
 import static org.eclipse.dataspaceconnector.registration.authority.model.ParticipantStatus.AUTHORIZED;
 import static org.eclipse.dataspaceconnector.registration.authority.model.ParticipantStatus.AUTHORIZING;
